[JIT] memory planning base with naive strategy
This PR adds static memory planning for traces with shape info. The three strategies implemented so far are stacked on top. They are:

* Linear scan (a heuristic based on https://www.usenix.org/legacy/events/vee05/full_papers/p132-wimmer.pdf)
* Greedy by size (a heuristic that allocates largest tensors first)
* Greedy by operator breadth (a heuristic that allocates largest tensors for operators with largest breadths first)

The latter 2 are based on https://arxiv.org/pdf/2001.03288.pdf.

Most of the action is under `torch/csrc/jit/passes/memory_planning/`.
Currently this only supports CPU. In principle there's nothing blocking supporting GPU (just requires adding cudaAllocator to GetAllocator) but there are lots of implicit assumptions encoded in the allocation strategies that are probably false for GPU.

Currently the tests don't do anything except exercise the code paths. As we find corner cases we will add tests to check for them.

ghstack-source-id: a6531eed1ad81a0622a308c24ba6fd7f319e26ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64347