OpInfos for some Tensor dtype conversion methods
OpInfos for:
- Tensor.bfloat16, Tensor.bool, Tensor.bypte, Tensor.char
- Tensor.double, Tensor.float, Tensor.half, Tensor.int
- Tensor.short, Tensor.long

None of these are supported by TorchScript. Also, the OpInfo autograd
test runner assumes that the operation is not allowed to change the
dtype of the argument, so only Tensor.double has
`supports_autograd=True` (in theory Tensor.bfloat16, Tensor.float,
Tensor.half should be differentiable).

Test Plan:
- run tests

ghstack-source-id: 3d3ceb8a10793985ac059b3690121b4458ed99ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64282