[Pytorch Edge] Support profiling kineto events from external source
Summary:
This diff exposes a way to add events to kineto profiler from external
source.
This can be a backend that executes a subgraph and wants to record this
execution in kineto profiler.
This diff also adds "backend" metadata to identify the backend an event
would have executed on.

Test Plan:
test_lite_interpreter

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 11526ece7b93cc6107c239865654a884f7a12d22
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64397