[quant][fx] Don't assume bias is a keyword argument
`prepare_fx` currently assumes that bias is always a positional argument to
convolutions, and only a key argument to other functions. This happens to work
today due to a quirk in how `__torch_function__` is handled for python
functions but shouldn't be considered stable.

Instead, we should support `bias` for both positional and keyword forms.

ghstack-source-id: 2fd9f60a7b6b9d6395a01391d15490b34499065a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61647