[sparsity] Remove the pack_param from the sparsifier state_dict
That was the original design, that we decided to simplify by removing the packing in the sparsifier.
The state of the sparsifier is saved directly, and the old behavior accidentally bled through to the current version.
This change removes the `_pack_params` method, and changes the state_dict to include the state directly.
We don't have to change the load_state_dict, as it will work with either the old or the new format.

ghstack-source-id: 52d5216591b9b891a1370423623694bba45674d1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62806